Restore config::update_config_with_includes; move included_config_scopes, scopes_from_paths to config
